backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
pydantic==2.5.0
pyyaml==6.0.1
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
eventlet==0.33.3
